+++ CHANGELOG.md
@@ -4,11 +4,21 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.3.0] - 2020-06-03
+
+## Added
+
+- Routes for accessing all tours related data.
+
+## Changed
+
+- README.md file.
+
 ## [0.2.0] - 2020-06-03
 
 ### Added
 
-- Basic express server with the base route (GET, '/').
+- Basic express server with the Base route.
 
 ### Changed